requirements.txt
Flask==1.1.2
requests==2.25.1
PyYAML
